fix compiler warnings related to signed/unsigned in packed fmatrix

+++ src/packed_fmatrix.hh
@@ -18,7 +18,7 @@
         __m256i mx = this->get(r0, col);                        \
         int mxi = r0;                                           \
         char cmpmsk = 1 << (VECTOR_N - 1 - index);              \
-        for (int row = r0 + 1; row < this->rows; row++)         \
+        for (uint64_t row = r0 + 1; row < this->rows; row++)    \
         {                                                       \
             char cmp = _mm256_cmp_epu32_mask(                   \
                 mx,                                             \
@@ -55,7 +55,7 @@
             mask,                                               \
             mask                                                \
         );                                                      \
-        for (int row = r0 + 1; row < this->rows; row++)         \
+        for (uint64_t row = r0 + 1; row < this->rows; row++)    \
         {                                                       \
             __m256i val = _mm256_permutevar8x32_epi32(          \
                 this->get(row, col),                            \
@@ -68,10 +68,10 @@
 class Packed_FMatrix
 {
 private:
-    int rows;
-    int cols;
+    uint64_t rows;
+    uint64_t cols;
     // original matrix n moduloe VECTOR_N
-    int nmod;
+    uint64_t nmod;
     std::vector<__m256i> m;
 
     __m256i get(int row, int col) const
@@ -117,7 +117,7 @@ public:
                 int c = this->cols - 1;
                 uint64_t elems[4];
                 elems[0] = 0; elems[1] = 0; elems[2] = 0; elems[3] = 0;
-                for (int i = 0; i < this->nmod; i++)
+                for (uint64_t i = 0; i < this->nmod; i++)
                     elems[i/2] |= matrix(r, VECTOR_N*c + i).get_repr() << (32*(1 - i%2));
 
                 this->set(r, c, _mm256_set_epi64x(
@@ -129,9 +129,9 @@ public:
                 );
             }
         }
-        for (int r = matrix.get_n(); r < this->rows; r++)
+        for (uint64_t r = matrix.get_n(); r < this->rows; r++)
         {
-            for (int c = 0; c < this->cols - 1; c++)
+            for (uint64_t c = 0; c < this->cols - 1; c++)
                 this->set(r, c, _mm256_setzero_si256());
 
             /* lazy.... */
@@ -228,7 +228,7 @@ public:
             0b111
         );
         /* first do left to right r1 */
-        for (int col = 0; col < this->cols; col++)
+        for (uint64_t col = 0; col < this->cols; col++)
         {
             /* already save them in reverse order here and permute,
              * values in reverse order, too*/
@@ -336,7 +336,7 @@ public:
                 break;
             }
 
-            for (int col = 0; col < this->cols - 1; col++)
+            for (uint64_t col = 0; col < this->cols - 1; col++)
                 coeffs[col] = _mm256_permutex2var_epi32(
                     coeffs[col],
                     idx,
@@ -439,7 +439,7 @@ public:
         }
 
         /* and do r2 left to right */
-        for (int col = 0; col < this->cols; col++)
+        for (uint64_t col = 0; col < this->cols; col++)
         {
             this->set(r2, col,
                       global::F.wide_mul(
@@ -453,7 +453,7 @@ public:
     void swap_rows(int r1, int r2)
     {
         __m256i tmp;
-        for (int c = 0; c < this->cols; c++)
+        for (uint64_t c = 0; c < this->cols; c++)
         {
             tmp = this->get(r1, c);
             this->set(r1, c, this->get(r2, c));
@@ -469,7 +469,7 @@ public:
             v << 32 | v,
             v << 32 | v
         );
-        for (int col = 0; col < this->cols; col++)
+        for (uint64_t col = 0; col < this->cols; col++)
             this->set(row, col,
                       global::F.wide_mul(this->get(row, col), pack)
                 );
@@ -478,7 +478,7 @@ public:
     /* subtract v times r1 from r2 */
     void row_op(int r1, int r2, __m256i pack)
     {
-        for (int col = 0; col < this->cols; col++)
+        for (uint64_t col = 0; col < this->cols; col++)
         {
             __m256i tmp = global::F.wide_mul(this->get(r1, col), pack);
 
@@ -491,7 +491,7 @@ public:
     GF_element det()
     {
         uint64_t det = 0x1;
-        for (int col = 0; col < this->cols; col++)
+        for (uint64_t col = 0; col < this->cols; col++)
         {
             DET_LOOP(0);
             DET_LOOP(1);
@@ -510,9 +510,9 @@ public:
     {
         std::valarray<GF_element> unpacked(this->rows * this->rows);
 
-        for (int row = 0; row < this->rows; row++)
+        for (uint64_t row = 0; row < this->rows; row++)
         {
-            for (int col = 0; col < this->cols; col++)
+            for (uint64_t col = 0; col < this->cols; col++)
             {
                 for (int e = 0; e < VECTOR_N; e++)
                 {
@@ -545,7 +545,7 @@ public:
         uint64_t n = this->rows;
         if (this->nmod)
             n -= VECTOR_N - this->nmod;
-        return FMatrix(n, unpacked[std::gslice(0, {n,n}, {this->rows,1})]);
+        return FMatrix(n, unpacked[std::gslice(0, {n, n}, {this->rows, 1})]);
     }
 };
